chore: renaming uuid gen module to generator
